requirements.txt
pandas 
mcp 
python-dotenv
langgraph
langchain-mcp-adapters
langchain-openai
fastapi
uvicorn[standard]
jinja2
